refactor:useCreateDecorateComponent 내부 객체 분리
* useCreateDecorateComponent 디렉토리 추가
   - properties.js 추가
